refactor: Make location-search module up-to-date with new coding guidelines
- Naming with correct suffixes
- Only accessed through index.ts
- Navigation/Routing types and helpers in its own file ´navigation-types.ts`
